Fix #794: Unify wallet and tipping imports across app/* and root lib/*
