+++ xconfess-frontend/app/lib/types/auth.ts
@@ -52,6 +52,7 @@ export interface AuthState {
   isAuthenticated: boolean;
   isLoading: boolean;
   error: string | null;
+  isSessionExpired: boolean;
 }
 
 /**
